+++ docs/unsafe-usage-baseline.md
@@ -8,7 +8,7 @@
 
 | module | files | unsafe | transmute | new_unchecked | assume_init | // SAFETY: |
 |---|---:|---:|---:|---:|---:|---:|
-| buffer | 7 | 60 | 0 | 0 | 0 | 9 |
+| buffer | 7 | 59 | 0 | 0 | 0 | 8 |
 | latch | 4 | 41 | 0 | 0 | 0 | 4 |
 | row | 3 | 6 | 0 | 0 | 1 | 3 |
 | index | 13 | 114 | 0 | 0 | 5 | 2 |
@@ -16,7 +16,7 @@
 | trx | 15 | 12 | 3 | 2 | 0 | 4 |
 | lwc | 2 | 8 | 6 | 0 | 3 | 0 |
 | file | 5 | 22 | 0 | 0 | 2 | 3 |
-| **total** | **52** | **309** | **11** | **2** | **11** | **28** |
+| **total** | **52** | **308** | **11** | **2** | **11** | **27** |
 
 ## File Hotspots (top 40)
 
@@ -25,26 +25,26 @@
 | `doradb-storage/src/index/btree_node.rs` | index | 53 | 0 |
 | `doradb-storage/src/io/mod.rs` | io | 30 | 3 |
 | `doradb-storage/src/index/btree.rs` | index | 21 | 2 |
-| `doradb-storage/src/buffer/guard.rs` | buffer | 18 | 4 |
 | `doradb-storage/src/latch/mutex.rs` | latch | 18 | 0 |
 | `doradb-storage/src/latch/rwlock.rs` | latch | 18 | 0 |
 | `doradb-storage/src/buffer/evict.rs` | buffer | 17 | 3 |
+| `doradb-storage/src/buffer/guard.rs` | buffer | 17 | 3 |
 | `doradb-storage/src/file/mod.rs` | file | 17 | 3 |
 | `doradb-storage/src/index/block_index.rs` | index | 16 | 0 |
 | `doradb-storage/src/index/column_block_index.rs` | index | 13 | 0 |
 | `doradb-storage/src/buffer/fixed.rs` | buffer | 12 | 0 |
 | `doradb-storage/src/io/libaio_abi.rs` | io | 12 | 0 |
-| `doradb-storage/src/buffer/util.rs` | buffer | 10 | 2 |
+| `doradb-storage/src/buffer/util.rs` | buffer | 8 | 0 |
 | `doradb-storage/src/trx/undo/row.rs` | trx | 7 | 4 |
 | `doradb-storage/src/index/btree_hint.rs` | index | 6 | 0 |
 | `doradb-storage/src/row/mod.rs` | row | 6 | 3 |
 | `doradb-storage/src/latch/hybrid.rs` | latch | 5 | 4 |
 | `doradb-storage/src/lwc/mod.rs` | lwc | 5 | 0 |
+| `doradb-storage/src/buffer/frame.rs` | buffer | 4 | 2 |
 | `doradb-storage/src/file/table_file.rs` | file | 4 | 0 |
 | `doradb-storage/src/io/buf.rs` | io | 4 | 0 |
 | `doradb-storage/src/lwc/page.rs` | lwc | 3 | 0 |
 | `doradb-storage/src/trx/redo.rs` | trx | 3 | 0 |
-| `doradb-storage/src/buffer/frame.rs` | buffer | 2 | 0 |
 | `doradb-storage/src/index/btree_scan.rs` | index | 2 | 0 |
 | `doradb-storage/src/index/unique_index.rs` | index | 2 | 0 |
 | `doradb-storage/src/buffer/page.rs` | buffer | 1 | 0 |
